fix: align implementation with openspec spec and add missing tests
Spec divergences fixed:
- GeoCircle: sealed record → readonly record struct
- GeoOrientation: sealed record → readonly record struct
- GeoCorridor.HalfWidthInMeters: double → int (per spec)
- Location: add missing AltitudeAccuracyInMeters property

Tests added:
- ExtendedGeographyTests (GeoCircle, GeoPolygon, GeoCorridor, GeoOrientation)
- GeometryTests (Angle, IntegerRange, Size2D, Point2D/3D, Anchor2D, Rectangle2D, AngleRange)
- LocationTests
- IdentifierTests (NameId, ExternalId)
- DurationTests
- ThreadSafetyReflectionTests (all public structs readonly, all record classes sealed)
- EnumOrdinalStabilityTests: UnitSystem, DayOfWeek, default(LocationAccuracy),
  default(EvseState), default(MapScheme), SpeedWarningStatus 2-member check,
  ChargingConnectorType.SaeJ3400 presence

132 tests passing (up from 66).

+++ src/Here.Sdk.Premium.Common/Geography/GeoCorridor.cs
@@ -9,11 +9,11 @@ public sealed record GeoCorridor
     public GeoPolyline Polyline { get; }
 
     /// <summary>Half-width in meters from the center-line. Must be &gt;= 0.</summary>
-    public double HalfWidthInMeters { get; }
+    public int HalfWidthInMeters { get; }
 
     /// <summary>Initializes a new <see cref="GeoCorridor"/>.</summary>
     /// <exception cref="ArgumentOutOfRangeException">When <paramref name="halfWidthInMeters"/> is negative.</exception>
-    public GeoCorridor(GeoPolyline polyline, double halfWidthInMeters)
+    public GeoCorridor(GeoPolyline polyline, int halfWidthInMeters)
     {
         if (halfWidthInMeters < 0)
             throw new ArgumentOutOfRangeException(nameof(halfWidthInMeters), "HalfWidthInMeters must be >= 0.");
